Chore: some basic code cleanup
